Make foreground controllable.
This deprecates the VNC option in favour of `FOREGROUND=vnc=:1`; in the
process it makes turning off the display (none mode), and sdl mode
possible.

For sdl mode, it tries to rely on DISPLAY and XAUTHORITY to work.  In
my local usage, this *does* work, but wouldn't surprise me if there are
other configurations where this doesn't (patches welcome).